Move service bus and telemetry to common lob to be used across all services
