+++ src/Clients/PullRequestQuantifier.GitHub.Client/Controllers/GitHubWebhookController.cs
@@ -9,10 +9,10 @@
     using Microsoft.Extensions.Options;
     using Newtonsoft.Json.Linq;
     using PullRequestQuantifier.Common;
-    using PullRequestQuantifier.GitHub.Client.Events;
+    using PullRequestQuantifier.Common.Azure.ServiceBus;
+    using PullRequestQuantifier.Common.Azure.Telemetry;
     using PullRequestQuantifier.GitHub.Client.GitHubClient;
     using PullRequestQuantifier.GitHub.Client.Models;
-    using PullRequestQuantifier.GitHub.Client.Telemetry;
 
     [ApiController]
     [Route("[controller]")]
