+++ src/Clients/PullRequestQuantifier.GitHub.Client/GitHubClient/GitHubClientMessageHandler.cs
@@ -4,7 +4,7 @@
     using System.Net.Http;
     using System.Threading;
     using System.Threading.Tasks;
-    using PullRequestQuantifier.GitHub.Client.Telemetry;
+    using PullRequestQuantifier.Common.Azure.Telemetry;
 
     public class GitHubClientMessageHandler : HttpClientHandler
     {
