+++ src/PullRequestQuantifier.Common/GlobalSuppressions.cs
@@ -1,0 +1,12 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Not applicable")]
+[assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1618:Generic type parameters should be documented", Justification = "Todo")]
+[assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1615:Element return value should be documented", Justification = "Todo")]
+[assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1611:Element parameters should be documented", Justification = "Todo")]
+[assembly: SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "_", Scope = "member", Target = "~M:PullRequestQuantifier.GitHub.Client.Events.GitHubEventHost.QuantifyPullRequest(System.String)~System.Threading.Tasks.Task")]
