+++ src/PullRequestQuantifier.Feedback.Service/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "PullRequestQuantifier.Feedback.Service": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "https://localhost:5001;http://localhost:5000"
+    }
+  }
+}
