+++ src/PullRequestQuantifier.GitEngine/Helpers/Diff/GitCreateDiff.cs
@@ -83,9 +83,7 @@
                     SetNormalAttribute(dirInfo);
                     dirInfo.Delete(true);
                 }
-#pragma warning disable CA1031 // Do not catch general exception types
                 catch
-#pragma warning restore CA1031 // Do not catch general exception types
                 {
                     // ignored
                 }
